refactor: drop bootstrap envs and merge signature helpers
