+++ docs/incident-runbook.md
@@ -11,10 +11,9 @@ When production goes sideways, use this guide to restore service quickly.
 
 ## Endpoint Failover
 
-1. **Update `EXPO_PUBLIC_WAKU_BOOTSTRAP`** (e.g., via `.env`) with a healthy list of peers if overriding defaults.
-2. **Override NEAR RPC:** restart deployments using `NEAR_RPC_URL=<backup-url>` or an appropriate flag.
-3. **Flush caches** with `expo start -c` to ensure new endpoints are used.
-4. **Monitor** connectivity logs from agents for successful peer discovery.
+1. **Override NEAR RPC:** restart deployments using `NEAR_RPC_URL=<backup-url>` or an appropriate flag.
+2. **Flush caches** with `expo start -c` to ensure new endpoints are used.
+3. **Monitor** connectivity logs from agents for successful peer discovery.
 
 ## Gateway Failover
