+++ services/__mocks__/nearSettings.ts
@@ -32,9 +32,6 @@ export async function fetchSettings() {
     rpcFallbackUrls: store['rpcFallbackUrls']
       ? JSON.parse(store['rpcFallbackUrls'])
       : [],
-    wakuBootstrap: store['wakuBootstrap']
-      ? JSON.parse(store['wakuBootstrap'])
-      : [],
     paymentFactoryAddress: store['paymentFactoryAddress'],
   };
 }
