--- utils/verifyBeforeWrite.ts
@@ -1,27 +1,0 @@
-import { z } from 'zod';
-import type { WakuMessage } from '../types/waku';
-import { errorLog } from './logger';
-import { verifyMessageSignature } from './verifyMessageSignature';
-
-export async function verifyBeforeWrite<T>(
-  data: unknown,
-  schema: z.ZodType<WakuMessage<T>>,
-  allowedPublicKeys?: string[],
-): Promise<WakuMessage<T> | null> {
-  const res = schema.safeParse(data);
-  if (!res.success) {
-    errorLog('Invalid Waku message');
-    return null;
-  }
-  const msg = res.data;
-  const ok = await verifyMessageSignature(msg, msg.sender.publicKey);
-  if (!ok) {
-    errorLog('E_SIGNATURE_INVALID');
-    return null;
-  }
-  if (allowedPublicKeys && !allowedPublicKeys.includes(msg.sender.publicKey)) {
-    errorLog('E_UNAUTHORIZED');
-    return null;
-  }
-  return msg;
-}
